[Scoper] Using different scoper.php for php 7.0
