+++ .github/workflows/build_scoped_rector_php70.yaml
@@ -44,7 +44,7 @@ jobs:
             -   run: sh build/downgrade-rector-php70.sh rector-build-php70
 
             # 3. prefix classes
-            -   run: sh build/build-rector-scoped.sh rector-build-php70 rector-prefixed-downgraded-php70
+            -   run: sh build/build-rector-scoped-php70.sh rector-build-php70 rector-prefixed-downgraded-php70
 
             # 4. lint the code for PHP 7.0 - this must happen here, as setup-php allows only one PHP version switch: https://github.com/shivammathur/setup-php/issues/434
             -
